Change: (hammy-define) Ensure that interval function slots are lists
i.e. lists of functions.  If a slot were, e.g. an uncompiled closure,
it would also be a list, but if a function symbol were pushed onto
that list, it would be neither a valid closure nor a list of
functions, so we force the slot to always be a list of functions (or
nil).