Change: (hammy--mode-line-update) Optimize
This way, for half the time, only one comparison is needed rather than
having to do all of them.

I suppose this assumes that a hammy will not spend most of its time
being overdue; were that the case, this could be counter-optimal.  Not
that it matters much, anyway, especially since we're about to avoid
excessive pie-making...