Add/Change: Avoid excessive pie-making
The hammy-mode-lighter function can be called a surprising number of
times.  For example, when calling (force-mode-line-update 'all), it
appears to happen twice, even though it's only displayed in the
tab-bar.  Also, when windows are scrolled, it can happen in rapid
succession.  So we don't want to be generating and rendering an SVG
every time this happens, especially when its value isn't
changing.

So now we cache the pie and the number of elapsed seconds in a hammy
when the pie was last updated, and we only update the pie when
necessary, according to this configured interval.  Ten seconds is
probably fine; for very short timers I suppose it could be inaccurate,
and users of such timers can adjust it.